+++ src/types/eventTypes.ts
@@ -1,5 +1,17 @@
 export type SensorEvent = {
-  id: string
-  coordinates: string
-  status: "registered" | "checkpoint" | "delayed"
-}
+  id: string;
+  publicKey: string;
+  name: string;
+  type: string;
+  lastEvent: {
+    lat: number;
+    lng: number;
+    status: string;
+  };
+};
+
+export type SensorEventDashboard = {
+  id: string;
+  coordinates: string;
+  status: string;
+};
